+++ packages/polaris-viz-core/src/types.ts
@@ -134,6 +134,7 @@ export interface TooltipTheme {
 }
 export interface SeriesColors {
   comparison: string;
+  empty: Color;
   single: Color;
   upToFour: Color[];
   fromFiveToSeven: Color[];